Implementation of circle drawing algorithm
- implement circle drawing algorithm,
- flip arguments for better readability of the default values and easier setup of the colour menu,
- move pixel settings (px_set) to the top of the stack.

The default pixel settings push onto the stack is there to avoid problems with variable assignments. Since rectangle takes 4 arguments and circle takes 3, if the pixel settings were pushed onto the stack first, they would be in different subsequent positions. Pushing px_set last ensures that in all cases it is first on the stack (bp + 4).