Optimise rectangle drawing algorithm
- rewrite the rectangle drawing algorithm to use STOSB instruction,
- update the boundaries tests to account for the changes,
- add unconditional jump in tests to save computation in certain cases,
- update the square dimensions in Vitruvian Man to account for the changes.

Previously, the algorithm relied on BIOS interrupt 10h that is slow by its nature -- it performs various tests before plotting a pixel on the screen. The new solution draws the rectangle directly in the video memory. It calculates the offset in the memory and sends a byte specifying a colour.

The outcome is a more performant way to draw a rectangle, but it still plots the pixels one by one, line after line. There is a way to draw 4, or even 8 pixels at a time. This is a potential improvement to explore in the future.