Add operand sizes to pushes, fix formatting
- add missing operand sizes to push instructions,
- fix indentation,
- adjust comments,
- remove redundant code,
- change radius on the default circle values,
- change radius of the circle option 4.

https://stackoverflow.com/a/45134007
The Stack Overflow comment linked above offers a lengthy explanation on push instruction in assembly. The user recommends not specifying the size of the operation and leaving it to the assembler, since it will use the appropriate size for the current mode and potentially optimise the operation. However, some assemblers need the size specified for operations on ambiguous values. Therefore I'm employing the good practice in the code and add missing operand sizes. Apparently, NASM optimises the operation despite providing the operand size. To enforce it, 'strict' keyword should be used along with the desired size.